move useFloatingUI from Popper to utils; remove Popper
