+++ packages/lab/src/dropdown/DropdownBase.tsx
@@ -5,7 +5,7 @@ import { Children, cloneElement, forwardRef, useRef, useState } from "react";
 import { forwardCallbackProps } from "../utils";
 import { DropdownBaseProps } from "./dropdownTypes";
 import { useDropdownBase } from "./useDropdownBase";
-import { useFloatingUI } from "../popper";
+import { useFloatingUI } from "../utils";
 import { Portal } from "../portal";
 import { isDesktop, useWindow } from "../window";
