+++ packages/lab/src/index.ts
@@ -52,7 +52,6 @@ export * from "./pagination";
 export * from "./parent-child-item";
 export * from "./parent-child-layout";
 export * from "./pill";
-export * from "./popper";
 export * from "./portal";
 export * from "./progress";
 export * from "./query-input";
