--- packages/lab/src/popper/index.ts
@@ -1,1 +1,0 @@
-export * from "./useFloatingUI";
