+++ packages/lab/src/utils/index.ts
@@ -6,6 +6,7 @@ export * from "./isPlainObject";
 export * from "./partition";
 export * from "./useClickOutside";
 export * from "./useEventCallback";
+export * from "./useFloatingUI";
 export * from "./useIsViewportLargerThanBreakpoint";
 export * from "./useLayoutEffectOnce";
 export * from "./useLayoutEffectSkipFirst";
